+++ src/SensirionI2CTxFrame.h
@@ -52,9 +52,9 @@ class SensirionI2CTxFrame {
     uint16_t addFloat(float data);
     uint16_t addBytes(uint8_t data[], size_t dataLength);
     uint16_t addCommand(uint16_t command);
-    uint16_t reset(void);
 
   private:
+    static uint8_t _generateCRC(const uint8_t* data, size_t count);
     uint16_t _addByte(uint8_t data);
     uint8_t* _buffer;
     size_t _bufferSize;